moved objectToParams helper to api module to avoid 'store' import in a helper file
