+++ app/src/helpers/commons.js
@@ -1,6 +1,3 @@
-import store from '@/store'
-
-
 /**
  * Allow to set a timeout on a `Promise` expected response.
  * The returned Promise will be rejected if the original Promise is not resolved or
@@ -16,31 +13,6 @@ export function timeout (promise, delay) {
     setTimeout(() => reject, delay)
     promise.then(resolve, reject)
   })
-}
-
-
-/**
- * Converts an object literal into an `URLSearchParams` that can be turned into a
- * query string or used as a body in a `fetch` call.
- *
- * @param {Object} obj - An object literal to convert.
- * @param {Object} options
- * @param {Boolean} [options.addLocale=false] - Option to append the locale to the query string.
- * @return {URLSearchParams}
- */
-export function objectToParams (obj, { addLocale = false } = {}) {
-  const urlParams = new URLSearchParams()
-  for (const [key, value] of Object.entries(obj)) {
-    if (Array.isArray(value)) {
-      value.forEach(v => urlParams.append(key, v))
-    } else {
-      urlParams.append(key, value)
-    }
-  }
-  if (addLocale) {
-    urlParams.append('locale', store.getters.locale)
-  }
-  return urlParams
 }
